requirements.txt
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
Django==5.0
djangorestframework==3.14.0
geographiclib==2.0
geopy==2.4.1
idna==3.6
Pillow==10.1.0
pytz==2023.3.post1
req==1.0.0
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
urllib3==2.1.0
